pkg/seq/ast: moves the structured AST nodes from doc package to inside ast package, replacing the previous content
