+++ pkg/sequence/ast/interfaces.go
@@ -1,0 +1,12 @@
+package ast
+
+type Stmt interface {
+	_stmt()
+}
+
+func (Alt) _stmt()            {}
+func (Break) _stmt()          {}
+func (CriticalRegion) _stmt() {}
+func (Loop) _stmt()           {}
+func (Message) _stmt()        {}
+func (Parallel) _stmt()       {}
